+++ src/pages/shop/ProductView/DetailViewPanel.tsx
@@ -2,6 +2,7 @@ import React, { useState } from "react"
 import { Divider, Button } from "@mui/material"
 import { DetailView, MoreButton, CoverBox, MoreInfomation } from "./detailViewPanel.styles"
 import { HalfArrow } from "../../../assets/images"
+import img_product_view from "../../../assets/images/img_product_view.png"
 
 const DetailViewPanel = () => {
   const [viewMore, setViewMore] = useState(false)
@@ -13,7 +14,7 @@ const DetailViewPanel = () => {
   return (
     <>
       <DetailView sx={{ height: viewMore ? 'auto' : '502px', pb: viewMore ? '58px' : 0 }}>
-        <img src="https://cdn.imweb.me/upload/S202012142df41e7544ce6/8f30f47c5ce2d.jpg" alt="" />
+        <img src={img_product_view} alt="" />
         <CoverBox>
           <MoreButton 
             variant="outlined"
