+++ src/pages/shop/ProductView/index.styles.ts
@@ -1,10 +1,6 @@
 import { styled } from "@mui/material/styles"
-import { Box, Button, FormControl } from "@mui/material"
+import { Box, Button } from "@mui/material"
 import { PrimaryButton } from "../../../styles/buttons.styles"
-
-interface ISelectProps {
-  isopen: string
-}
 
 export const ProductMainInfo = styled(Box)(() => ({
   '& > .product_brand': {
@@ -114,83 +110,6 @@ export const ViewTitle = styled(Box)(() => ({
   '& > span': {
     color: '#868686',
     fontWeight: 'normal',
-  },
-}))
-
-export const OptionSelectType = styled(Box)(() => ({
-  '& .MuiInputBase-root': {
-    borderRadius: 0,
-  },
-  '& .MuiSelect-select': {
-    padding: '14.5px 14px',
-    fontSize: '.875rem',
-  },
-}))
-
-export const OutlinedSelect = styled(Box)<ISelectProps>((props) => ({
-  position: 'relative',
-  '&::before': {
-    content: '"⌵"',
-    position: 'absolute',
-    top: '8px',
-    right: '16px',
-    color: '#DADADA',
-    fontSize: '20px',
-  },
-  '& > .MuiButton-root': {
-    padding: '6px 16px',
-    minHeight: 52,
-    borderRadius: 0,
-    border: '1px solid #DADADA',
-    justifyContent: 'start',
-    color: '#999'
-  },
-  '& > .MuiButton-root > img': {
-    width: '18px',
-    marginRight: '4px',
-  },
-  '& .dropmenu': {
-    display: props.isopen === 'true' ? 'block' : 'none',
-    marginTop: '-1px',
-    padding: 0,
-    color:'#999999',
-    border: '1px solid #dadada',
-  },
-  '& .dropmenu > li': {
-    padding: 0,
-  },
-  '& .dropmenu > li:not(:last-of-type)': {
-    borderBottom: '1px solid #dadada',
-  },
-  '& .dropmenu .MuiListItemButton-root': {
-    padding: '13.5px 16px',
-  },
-  '& .dropmenu .MuiListItemButton-root > img': {
-    width: '18px',
-    marginRight: '4px',
-  }
-}))
-
-export const OutlinedCheckbox = styled(FormControl)(() => ({
-  '& input': {
-    display: 'none',
-  },
-  '& label': {
-    display: 'flex',
-    alignItems: 'center',
-    padding: '6.5px 12px',
-    borderRadius: '18px',
-    border: '1px solid #DADADA',
-    fontSize: '0.875rem',
-  },
-  '& input:checked + label': {
-    border: '1px solid #242223',
-    backgroundColor: '#242223',
-    color: '#fff',
-  },
-  '& label > img': {
-    width: '18px',
-    marginRight: '4px',
   },
 }))
