+++ src/pages/shop/WriteQna.tsx
@@ -1,0 +1,15 @@
+import React, { useEffect, useState } from "react"
+import { Tabs, Tab, Box, List, ListItem, ListItemButton, ListItemText } from "@mui/material"
+import CenterTitle from "../../components/title/CenterTitle"
+
+
+const WriteQna = () => {
+
+  return (
+    <>
+      <CenterTitle title={['문의하기']} />
+    </>
+  )
+}
+
+export default WriteQna
